ajusts in labels #2
