+++ src/components/GenericTable/EnhancedTableToolbar.tsx
@@ -1,0 +1,56 @@
+import { IconButton, Toolbar, Tooltip, Typography } from '@material-ui/core';
+import {
+  createStyles,
+  lighten,
+  makeStyles,
+  Theme,
+} from '@material-ui/core/styles';
+import FilterListIcon from '@material-ui/icons/FilterList';
+import { EnhancedTableToolbarProps } from './types';
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    root: {
+      paddingLeft: theme.spacing(2),
+      paddingRight: theme.spacing(1),
+    },
+    highlight:
+      theme.palette.type === 'light'
+        ? {
+          color: theme.palette.secondary.main,
+          backgroundColor: lighten(theme.palette.secondary.light, 0.85),
+        }
+        : {
+          color: theme.palette.text.primary,
+          backgroundColor: theme.palette.secondary.dark,
+        },
+    title: {
+      flex: '1 1 100%',
+    },
+  })
+);
+
+const EnhancedTableToolbar = (props: EnhancedTableToolbarProps) => {
+  const classes = useStyles();
+  const { title } = props;
+
+  return (
+    <Toolbar className={classes.root}>
+      <Typography
+        className={classes.title}
+        variant="h6"
+        id="tableTitle"
+        component="div"
+      >
+        {title}
+      </Typography>
+      <Tooltip title="Filter list">
+        <IconButton aria-label="filter list">
+          <FilterListIcon />
+        </IconButton>
+      </Tooltip>
+    </Toolbar>
+  );
+};
+
+export default EnhancedTableToolbar;
